status message 関数の集約
