+++ srcs/Server/HttpResponse.cpp
@@ -5,6 +5,47 @@
 #include "Utils.hpp"
 #include "sys/socket.h"
 
+namespace http {
+std::string GetStatusMessage(const int error_code) {
+  switch (error_code) {
+    case 200:
+      return "OK";
+    case 201:
+      return "Created";
+    case 202:
+      return "Accepted";
+    case 204:
+      return "No Content";
+    case 301:
+      return "Moved Permanently";
+    case 302:
+      return "Found";
+    case 304:
+      return "Not Modified";
+    case 307:
+      return "Temporary Redirect";
+    case 400:
+      return "Bad Request";
+    case 403:
+      return "Forbidden";
+    case 404:
+      return "Not Found";
+    case 405:
+      return "Method Not Allowed";
+    case 413:
+      return "Payload Too Large";
+    case 500:
+      return "Internal Server Error";
+    case 501:
+      return "Not Implemented";
+    case 505:
+      return "HTTP Version Not Supported";
+    default:
+      return "";
+  }
+}
+}  // namespace http
+
 const char HttpResponse::kCRLF[] = "\r\n";
 
 HttpResponse::HttpResponse()
